fix circular deps in editor
